backend/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.7.1
flake8==3.9.2
Markdown==3.3.4
mccabe==0.6.1
mixer==7.1.2
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
drf-yasg==1.20.0
psycopg2==2.8.6